requirements.txt
accelerate==0.33.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
applang==0.2.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
compressed-tensors==0.6.0
ConstructiveMath==0.1
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
Deprecated==1.2.18
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
easydict==1.13
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
gguf==0.10.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
interegular==0.3.3
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonargparse==4.39.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lark==1.2.2
libcst==1.7.0
litellm==1.69.2
llvmlite==0.44.0
lm-format-enforcer==0.10.6
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mistral_common==1.5.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
openai==1.75.0
opencv-python-headless==4.11.0.86
outlines==0.0.46
packaging==25.0
pandas==1.4.3
partial-json-parser==0.2.1.1.post5
pendulum==3.1.0
pillow==11.2.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==20.0.0
pycountry==24.6.1
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2022.1
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
roman==5.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.6.0
tiktoken==0.9.0
together==1.5.8
tokenizers==0.21.1
toml==0.10.2
torch==2.4.0
torchvision==0.19.0
tqdm==4.67.1
transformers==4.51.3
triton==3.0.0
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vllm==0.6.3.post1
vllm-nccl-cu12==2.18.1.0.4.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
